+++ src/base/static/js/routes.js
@@ -20,6 +20,7 @@ Shareabouts.Util = Util;
       ':dataset/:id': 'viewPlace',
       'new': 'newPlace',
       ':dataset/:id/response/:response_id': 'viewPlace',
+      ':id/response/:response_id': 'viewLandmark',
       ':dataset/:id/edit': 'editPlace',
       'list': 'showList',
       ':id': 'viewLandmark',
@@ -165,12 +166,21 @@ Shareabouts.Util = Util;
       this.appView.newPlace();
     },
 
-    viewLandmark: function(id) {
-      this.appView.viewLandmark(id, { zoom: this.loading });
+    viewLandmark: function(modelId, responseId) {
+      this.appView.viewPlaceOrLandmark({
+        modelId: modelId,
+        responseId: responseId,
+        loading: this.loading
+      });
     },
 
-    viewPlace: function(datasetSlug, id, responseId) {
-      this.appView.viewPlace(datasetSlug, id, responseId, this.loading);
+    viewPlace: function(datasetSlug, modelId, responseId) {      
+      this.appView.viewPlaceOrLandmark({
+        datasetSlug: datasetSlug,
+        modelId: modelId,
+        responseId: responseId,
+        loading: this.loading
+      });
     },
 
     editPlace: function(){},