+++ src/base/static/js/views/layer-view.js
@@ -1,4 +1,5 @@
   var Util = require('../utils.js');
+  var Argo = require('../../libs/leaflet.argo.js');
 
   module.exports = Backbone.View.extend({
      // A view responsible for the representation of a place on the map.
@@ -106,9 +107,13 @@
     },
     onMarkerClick: function() {
       Util.log('USER', 'map', 'place-marker-click', this.model.getLoggingDetails());
-      this.options.router.navigate('/' + this.model.get('datasetSlug') + '/' + this.model.id, {trigger: true});
+      // support places with landmark-style urls
+      if (this.model.get("url-title")) {
+        this.options.router.navigate('/' + this.model.get("url-title"), {trigger: true});
+      } else {
+        this.options.router.navigate('/' + this.model.get('datasetSlug') + '/' + this.model.id, {trigger: true});
+      }      
     },
-
     isPoint: function() {
       return this.model.get('geometry').type == 'Point';
     },