+++ src/base/static/js/views/place-list-view.js
@@ -48,9 +48,11 @@
     onRender: function(evt) {
       this.support.reset();
       this.support.show(this.supportView);
-      // in case story mode has hidden the title
+      
+      // in case story mode has hidden the title or inserted navigation bars
       if (this.model.get("story")) {
         this.$el.find(".place-header-title").removeClass("is-visuallyhidden");
+        this.$el.find(".place-story-bar").remove();
       }
     },
     show: function() {
@@ -78,7 +80,8 @@
       'submit @ui.searchForm': 'handleSearchSubmit',
       'click @ui.date': 'handleDateSort',
       'click @ui.surveyCount': 'handleSurveyCountSort',
-      'click @ui.supportCount': 'handleSupportCountSort'
+      'click @ui.supportCount': 'handleSupportCountSort',
+      'scroll': 'infiniteScroll'
     },
     initialize: function(options) {
       var self = this;
@@ -88,10 +91,14 @@
       // merged together, for sorting and filtering purposes
       this.collection = new PlaceCollection([]);
 
+      this.unrenderedItems = new PlaceCollection([]);
+
       _.each(this.options.placeCollections, function(collection) {
         collection.on("add", self.addModel, self);
-        collection.on("sync", self.onSync, self);
       });
+
+      this.itemsPerPage = 10;
+      this.numItemsShown = this.itemsPerPage;
 
       // Init the views cache
       this.views = {};
@@ -103,17 +110,16 @@
       this.collectionFilters = options.filter || {};
       this.searchTerm = options.term || '';
     },
-    onSync: function() {
-      // sort the merged collection after each component collection
-      // is synced successfully
-      this.sort();
-    },
     onAfterItemAdded: function(view) {
       // Cache the views as they are added
       this.views[view.model.cid] = view;
     },
     addModel: function(model) {
-      this.collection.add(model);
+      if (this.collection.length < this.numItemsShown) {
+        this.collection.add(model, {sort: false});
+      } else {
+        this.unrenderedItems.add(model, {sort: false});
+      }
     },
     renderList: function() {
       var self = this;
@@ -123,27 +129,36 @@
       var $itemViewContainer = this.getItemViewContainer(this);
       $itemViewContainer.empty();
 
-      this.collection.each(function(model) {
-        if (self.views[model.cid]) {
+      this.collection.each(function(model, index) {
+        if (self.views[model.cid] && index < self.numItemsShown) {
           $itemViewContainer.append(self.views[model.cid].$el);
           // Delegate the events so that the subviews still work
           self.views[model.cid].supportView.delegateEvents();
         }
       });
-
-      // remove story bars from the list view
-      $("#list-container .place-story-bar").remove();
+    },
+    infiniteScroll: function() {
+      var totalHeight = this.$('> ul').height();
+      var scrollTop = this.$el.scrollTop() + this.$el.height();
+      // 200 = number of pixels from bottom to load more
+      if (scrollTop + 200 >= totalHeight) {
+        this.numItemsShown += this.itemsPerPage;
+        this.applyFilters(this.collectionFilters, this.searchTerm, this.numItemsShown);
+      }
     },
     handleSearchInput: function(evt) {
       evt.preventDefault();
+      this.numItemsShown = this.itemsPerPage;
       this.search(this.ui.searchField.val());
     },
     handleSearchSubmit: function(evt) {
       evt.preventDefault();
+      this.numItemsShown = this.itemsPerPage;
       this.search(this.ui.searchField.val());
     },
     handleDateSort: function(evt) {
       evt.preventDefault();
+      this.numItemsShown = this.itemsPerPage;
       this.sortBy = 'date';
       this.sort();
 
@@ -151,6 +166,7 @@
     },
     handleSurveyCountSort: function(evt) {
       evt.preventDefault();
+      this.numItemsShown = this.itemsPerPage;
       this.sortBy = 'surveyCount';
       this.sort();
 
@@ -158,6 +174,7 @@
     },
     handleSupportCountSort: function(evt) {
       evt.preventDefault();
+      this.numItemsShown = this.itemsPerPage;
       this.sortBy = 'supportCount';
       this.sort();
 
@@ -214,50 +231,48 @@
       var sortFunction = this.sortBy + 'Sort';
 
       this.collection.comparator = this[sortFunction];
+      this.unrenderedItems.comparator = this[sortFunction];
       this.collection.sort();
+      this.unrenderedItems.sort();
       this.renderList();
       this.search(this.ui.searchField.val());
     },
     clearFilters: function() {
       this.collectionFilters = {};
-      this.applyFilters(this.collectionFilters, this.searchTerm);
+      this.applyFilters(this.collectionFilters, this.searchTerm, this.numItemsShown);
     },
     filter: function(filters) {
       _.extend(this.collectionFilters, filters);
-      this.applyFilters(this.collectionFilters, this.searchTerm);
+      this.applyFilters(this.collectionFilters, this.searchTerm, this.numItemsShown);
     },
     search: function(term) {
       this.searchTerm = term;
-      this.applyFilters(this.collectionFilters, this.searchTerm);
+      this.applyFilters(this.collectionFilters, this.searchTerm, this.numItemsShown);
     },
-    applyFilters: function(filters, term) {
-      var val, key, i;
+    applyFilters: function(filters, term, max) {
+      var val, key, i, 
+          previouslyUnrenderedModels = [];
 
       term = term.toUpperCase();
-      this.collection.each(function(model) {
-        var show = function() { model.trigger('show'); },
-            hide = function() { model.trigger('hide'); },
-            submitter, 
+      this.unrenderedItems.add(this.collection.models);
+      this.collection.reset();
+
+      this.unrenderedItems.each(function(model, index) {
+        if (index > max) {
+          return;
+        }
+
+        var submitter,
             locationType = model.get("location_type"),
             placeConfig = _.find(Shareabouts.Config.place.place_detail, function(config) { return config.category === locationType });
 
-        // If the model doesn't match one of the filters, hide it.
-        for (key in filters) {
-          val = filters[key];
-          if (_.isFunction(val) && !val(model)) {
-            return hide();
-          }
-          else if (!model.get(key) || val.toUpperCase() !== model.get(key).toUpperCase()) {
-            return hide();
-          }
-        }
-
         // Check whether the remaining models match the search term
-        for (var i = 0; i < placeConfig.fields.length; i++) { 
+        for (var i = 0; i < placeConfig.fields.length; i++) {
           key = placeConfig.fields[i].name;
           val = model.get(key);
           if (_.isString(val) && val.toUpperCase().indexOf(term) !== -1) {
-            return show();
+            previouslyUnrenderedModels.push(model);
+            return this.collection.add(model);
           }
         };
 
@@ -265,23 +280,26 @@
         // with FB or Twitter. We handle it specially because it is an object,
         // not a string.
         submitter = model.get('submitter');
-        if (!show && submitter) {
+        if (submitter) {
           if (submitter.name && submitter.name.toUpperCase().indexOf(term) !== -1 ||
               submitter.username && submitter.username.toUpperCase().indexOf(term) !== -1) {
-            return show();
+            previouslyUnrenderedModels.push(model);
+            return this.collection.add(model);
           }
         }
 
         // If the location_type has a label, we should search in it also.
         locationType = Shareabouts.Config.flavor.place_types[model.get('location_type')];
-        if (!show && locationType && locationType.label) {
+        if (locationType && locationType.label) {
           if (locationType.label.toUpperCase().indexOf(term) !== -1) {
-            return show();
+            previouslyUnrenderedModels.push(model);
+            return this.collection.add(model);
           }
         }
+      }, this);
 
-        // If we've fallen through here, hide the item.
-        return hide();
+      previouslyUnrenderedModels.forEach(function(model) {
+        this.unrenderedItems.remove(model);
       }, this);
     },
     isVisible: function() {